requirements.txt
alembic==1.0.11
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.23
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
msrest==0.6.9
oauthlib==3.1.0
psycopg2==2.8.3
pyasn1==0.4.6
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.7
urllib3==1.25.3
Werkzeug==0.15.5
